@@ -5,7 +5,7 @@ homepage: https://github.com/ConnectyCube/connectycube-flutter-call-kit
 issue_tracker: https://github.com/ConnectyCube/connectycube-flutter-call-kit/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
